contract integration test

+++ ethereum-ra-api/src/main/java/be/e_contract/ethereum/ra/api/EthereumConnection.java
@@ -164,7 +164,7 @@ public interface EthereumConnection extends Connection, AutoCloseable {
      * @throws be.e_contract.ethereum.ra.api.EthereumException
      */
     TransactionReceipt deploy(Class<? extends Contract> contractClass, BigInteger gasPrice,
-            BigInteger gasLimit, Credentials credentials, Byte chainId) throws ResourceException, EthereumException;
+            BigInteger gasLimit, Credentials credentials, Integer chainId) throws ResourceException, EthereumException;
 
     /**
      * Loads a contract.
@@ -180,7 +180,7 @@ public interface EthereumConnection extends Connection, AutoCloseable {
      * @throws ResourceException
      */
     <T extends Contract> T load(Class<T> contractClass, String contractAddress,
-            Credentials credentials, Byte chainId, BigInteger gasPrice, BigInteger gasLimit) throws ResourceException;
+            Credentials credentials, Integer chainId, BigInteger gasPrice, BigInteger gasLimit) throws ResourceException;
 
     /**
      * Retrieve the chain identifier of the Ethereum network.
